Add the software update logout script
This script is to be used with JAMF's casper suite. It is part of a
smart group when JSS sees that the computer needs an update. This will
then prompt the user about installing the updates, they can click to
not install the updates 4 times before it just will install them.